requirements.txt
fastapi==0.116.1
jose==1.0.0
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
SQLAlchemy==2.0.43
